double vs int in an eternal fight

+++ cnF2freq.cpp
@@ -2686,7 +2686,7 @@ template<int N> struct valuereporter
 
 struct genotypereporter : valuereporter<4>
 {
-	void addval(int q, int mapval, int g, int flag2, int val)
+	void addval(int q, int mapval, int g, int flag2, double val)
 	{
 		probs[mapval] += val;
 	}
